+++ changelog.md
@@ -6,3 +6,11 @@
 [v0.3] Restore Windows File Explorer windows after restarting the process.  
 [v0.2] Added the `-Permanent` switch, which creates a scheduled task to automatically reapply the patch on every restart.  
 [v0.1] Initial version
+
+## Interactive Layer
+
+[v0.1] Interactive layer
+
+## Launcher (Launch Interactive.cmd)
+
+[v0.1] Simple PowerShell script launcher.